Actions/Closures as context parameters
Resolves #22

Adds support for `context.parameter(...)` to return a closure for use in action-type scenarios.

The debug view shows the name and signature of the closure, along with a count of the times the closure has been executed.

On execution, the debug log will show the name along with the parameters passed into the execution.

`CustomButton` has been modified to pass a `Date` in its action closure as an example.

Currently without variadic generics, we must manually implement the various parameter counts. At the moment, I've included up to 4 parameters. If more are needed, the types are public such that a consumer can extend `Exhibit.Context` to add additional cases.