Add custom layout builder to Exhibit (#27)
Resolves #4 

Allows for customization of the view that is displayed in the exhibition. For example, adding padding, inserting in a list, positioning, etc.

This layout builder does not affect SwiftUI previews.

Also uses `sizeThatFits` as default preview layout. Consumers can implement their own `static var previews` if they wish to change this

![Screen Shot 2022-02-23 at 12 22 59](https://user-images.githubusercontent.com/609274/155401962-efff925d-549e-4f88-95cb-4a28f025065d.png)