Exploration into custom presentation via Layout
Using the `layout` system for presentation has exposed a large issue with the layout system design, and perhaps with the underlying exhibit Context -> Debug menu relationship.

The layout closure allows for insertion of additional views before conditionally displaying the exhibit.

This breaks the preference relationship that passes the Context to the debug menu.

Resolving this could involve refactoring the Context passing mechanism, or adjusting the layout system to work with the current mechanism. Further exploration is needed.