Support modifying parameters in multiple previews
Resolves #14

This allows for consumers to override parameters in previews by using `Exhibit.preview(parameters: [String: Any]) -> some View`

This allows for multiple previews to be had for easing development.

In order to support the passing of parameters in this way, some refactoring was required:

`Exhibit` is now a strict model, not a view itself, allowing it to be generic and more easily understood.

`AnyExhibit` is added which is a type-erased version for use in arrays, and then `AnyExhibitView` displays the actual exhibit with support for context observing.

For some reason, a compiler error unfortunately occurs in the preview definition if `layout` is left as a closure on `Exhibit`. The only solution I could find which left types intact is to move the layout override to a separate function in the `Exhibit` protocol.